docs(nvim): suppress some diags in netrw.lua
This one I'm suppressing is for logical reasons.
The check before it checks whether the thing is a symlink. If it is then
we call fs_readlink. Because the original file is a symlink, calling
readlink on it the first time (which is what the variable `dest` is)
always succeed. So, dest can never be `nil`.

Commit-by: Huy Nguyen