refactor(nvim): more config tweaks
Treesitter context with the context on the cursor. After all, that's
where I'm editing, so context at the cursor is more useful than that of
the top line.
Remove the move-to @_start and @_end. There are rules for moving with
these, but it often takes some brain juice to think about. Not fast. And
the move-to @_start (which I map with ]b and [b (and capitalized)) are
in conflict with one of my frequently-used Mini keybinds (the next-prev
buffers configured by mini.bracketed (I believe)).

Disable ";" of mini.jump. I map that to treesitter-context's repeat
action. The latter requires like 3 or 4 keystrokes, while the former,
I can just mash "f" or "t".

Commit-by: Huy Nguyen